+++ src/DynamicExpresso.Core/Interpreter.cs
@@ -1,7 +1,9 @@
 using DynamicExpresso.Parsing;
 using DynamicExpresso.Reflection;
+using DynamicExpresso.Visitors;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Linq.Expressions;
 using DynamicExpresso.Exceptions;
 
@@ -13,14 +15,15 @@ namespace DynamicExpresso
 	/// </summary>
 	public class Interpreter
 	{
-		private readonly InterpreterSettings _settings;
+		private readonly ParserSettings _settings;
+		private readonly ISet<ExpressionVisitor> _visitors = new HashSet<ExpressionVisitor>();
 
 		#region Constructors
-
 		/// <summary>
 		/// Creates a new Interpreter using InterpreterOptions.Default.
 		/// </summary>
-		public Interpreter() : this(InterpreterOptions.Default)
+		public Interpreter()
+			: this(InterpreterOptions.Default)
 		{
 		}
 
@@ -30,7 +33,34 @@ namespace DynamicExpresso
 		/// <param name="options"></param>
 		public Interpreter(InterpreterOptions options)
 		{
-			_settings = new InterpreterSettings(options);
+			var caseInsensitive = options.HasFlag(InterpreterOptions.CaseInsensitive);
+
+			var lateBindObject = options.HasFlag(InterpreterOptions.LateBindObject);
+
+			_settings = new ParserSettings(caseInsensitive, lateBindObject);
+
+			if ((options & InterpreterOptions.SystemKeywords) == InterpreterOptions.SystemKeywords)
+			{
+				SetIdentifiers(LanguageConstants.Literals);
+			}
+
+			if ((options & InterpreterOptions.PrimitiveTypes) == InterpreterOptions.PrimitiveTypes)
+			{
+				Reference(LanguageConstants.PrimitiveTypes);
+				Reference(LanguageConstants.CSharpPrimitiveTypes);
+			}
+
+			if ((options & InterpreterOptions.CommonTypes) == InterpreterOptions.CommonTypes)
+			{
+				Reference(LanguageConstants.CommonTypes);
+			}
+
+			if ((options & InterpreterOptions.LambdaExpressions) == InterpreterOptions.LambdaExpressions)
+			{
+				_settings.LambdaExpressions = true;
+			}
+
+			_visitors.Add(new DisableReflectionVisitor());
 		}
 
 		/// <summary>
@@ -38,30 +68,52 @@ namespace DynamicExpresso
 		/// </summary>
 		internal Interpreter(ParserSettings settings)
 		{
-			_settings = new InterpreterSettings(settings);
+			_settings = settings;
 		}
-
 		#endregion
 
 		#region Properties
-
-		public bool CaseInsensitive { get { return _settings.CaseInsensitive; } }
+		public bool CaseInsensitive
+		{
+			get
+			{
+				return _settings.CaseInsensitive;
+			}
+		}
 
 		/// <summary>
 		/// Gets a list of registeres types. Add types by using the Reference method.
 		/// </summary>
-		public IEnumerable<ReferenceType> ReferencedTypes { get { return _settings.ReferencedTypes; } }
+		public IEnumerable<ReferenceType> ReferencedTypes
+		{
+			get
+			{
+				return _settings.KnownTypes
+					.Select(p => p.Value)
+					.ToList();
+			}
+		}
 
 		/// <summary>
 		/// Gets a list of known identifiers. Add identifiers using SetVariable, SetFunction or SetExpression methods.
 		/// </summary>
-		public IEnumerable<Identifier> Identifiers { get { return _settings.Identifiers; } }
+		public IEnumerable<Identifier> Identifiers
+		{
+			get
+			{
+				return _settings.Identifiers
+					.Select(p => p.Value)
+					.ToList();
+			}
+		}
 
 		/// <summary>
 		/// Gets the available assignment operators.
 		/// </summary>
-		public AssignmentOperators AssignmentOperators { get { return _settings.AssignmentOperators; } }
-
+		public AssignmentOperators AssignmentOperators
+		{
+			get { return _settings.AssignmentOperators; }
+		}
 		#endregion
 
 		#region Options
@@ -73,26 +125,29 @@ namespace DynamicExpresso
 		/// <returns></returns>
 		public Interpreter SetDefaultNumberType(DefaultNumberType defaultNumberType)
 		{
-			_settings.SetDefaultNumberType(defaultNumberType);
+			_settings.DefaultNumberType = defaultNumberType;
 			return this;
 		}
 
 		/// <summary>
-		/// Allows to enable/disable assignment operators.
+		/// Allows to enable/disable assignment operators. 
 		/// For security when expression are generated by the users is more safe to disable assignment operators.
 		/// </summary>
 		/// <param name="assignmentOperators"></param>
 		/// <returns></returns>
 		public Interpreter EnableAssignment(AssignmentOperators assignmentOperators)
 		{
-			_settings.EnableAssignment(assignmentOperators);
+			_settings.AssignmentOperators = assignmentOperators;
+
 			return this;
 		}
 		#endregion
 
 		#region Visitors
-
-		public ISet<ExpressionVisitor> Visitors { get { return _settings.Visitors; } }
+		public ISet<ExpressionVisitor> Visitors
+		{
+			get { return _visitors; }
+		}
 
 		/// <summary>
 		/// Enable reflection expression (like x.GetType().GetMethod() or typeof(double).Assembly) by removing the DisableReflectionVisitor.
@@ -100,14 +155,15 @@ namespace DynamicExpresso
 		/// <returns></returns>
 		public Interpreter EnableReflection()
 		{
-			_settings.EnableReflection();
+			var visitor = Visitors.FirstOrDefault(p => p is DisableReflectionVisitor);
+			if (visitor != null)
+				Visitors.Remove(visitor);
+
 			return this;
 		}
-
 		#endregion
 
 		#region Register identifiers
-
 		/// <summary>
 		/// Allow the specified function delegate to be called from a parsed expression.
 		/// </summary>
@@ -116,7 +172,18 @@ namespace DynamicExpresso
 		/// <returns></returns>
 		public Interpreter SetFunction(string name, Delegate value)
 		{
-			_settings.SetFunction(name, value);
+			if (string.IsNullOrWhiteSpace(name))
+				throw new ArgumentNullException(nameof(name));
+
+			if (_settings.Identifiers.TryGetValue(name, out var identifier) && identifier is FunctionIdentifier fIdentifier)
+			{
+				fIdentifier.AddOverload(value);
+			}
+			else
+			{
+				SetIdentifier(new FunctionIdentifier(name, value));
+			}
+
 			return this;
 		}
 
@@ -128,8 +195,10 @@ namespace DynamicExpresso
 		/// <returns></returns>
 		public Interpreter SetVariable(string name, object value)
 		{
-			_settings.SetVariable(name, value);
-			return this;
+			if (string.IsNullOrWhiteSpace(name))
+				throw new ArgumentNullException(nameof(name));
+
+			return SetExpression(name, Expression.Constant(value));
 		}
 
 		/// <summary>
@@ -140,8 +209,7 @@ namespace DynamicExpresso
 		/// <returns></returns>
 		public Interpreter SetVariable<T>(string name, T value)
 		{
-			_settings.SetVariable(name, value);
-			return this;
+			return SetVariable(name, value, typeof(T));
 		}
 
 		/// <summary>
@@ -153,8 +221,12 @@ namespace DynamicExpresso
 		/// <returns></returns>
 		public Interpreter SetVariable(string name, object value, Type type)
 		{
-			_settings.SetVariable(name, value, type);
-			return this;
+			if (type == null)
+				throw new ArgumentNullException(nameof(type));
+			if (string.IsNullOrWhiteSpace(name))
+				throw new ArgumentNullException(nameof(name));
+
+			return SetExpression(name, Expression.Constant(value, type));
 		}
 
 		/// <summary>
@@ -166,8 +238,7 @@ namespace DynamicExpresso
 		/// <returns></returns>
 		public Interpreter SetExpression(string name, Expression expression)
 		{
-			_settings.SetExpression(name, expression);
-			return this;
+			return SetIdentifier(new Identifier(name, expression));
 		}
 
 		/// <summary>
@@ -178,7 +249,9 @@ namespace DynamicExpresso
 		/// <returns></returns>
 		public Interpreter SetIdentifiers(IEnumerable<Identifier> identifiers)
 		{
-			_settings.SetIdentifiers(identifiers);
+			foreach (var i in identifiers)
+				SetIdentifier(i);
+
 			return this;
 		}
 
@@ -190,7 +263,14 @@ namespace DynamicExpresso
 		/// <returns></returns>
 		public Interpreter SetIdentifier(Identifier identifier)
 		{
-			_settings.SetIdentifier(identifier);
+			if (identifier == null)
+				throw new ArgumentNullException(nameof(identifier));
+
+			if (LanguageConstants.ReservedKeywords.Contains(identifier.Name))
+				throw new InvalidOperationException($"{identifier.Name} is a reserved word");
+
+			_settings.Identifiers[identifier.Name] = identifier;
+
 			return this;
 		}
 		#endregion
@@ -204,8 +284,10 @@ namespace DynamicExpresso
 		/// <returns></returns>
 		public Interpreter Reference(Type type)
 		{
-			_settings.Reference(type);
-			return this;
+			if (type == null)
+				throw new ArgumentNullException(nameof(type));
+
+			return Reference(type, type.Name);
 		}
 
 		/// <summary>
@@ -216,7 +298,12 @@ namespace DynamicExpresso
 		/// <returns></returns>
 		public Interpreter Reference(IEnumerable<ReferenceType> types)
 		{
-			_settings.Reference(types);
+			if (types == null)
+				throw new ArgumentNullException(nameof(types));
+
+			foreach (var t in types)
+				Reference(t);
+
 			return this;
 		}
 
@@ -229,8 +316,7 @@ namespace DynamicExpresso
 		/// <returns></returns>
 		public Interpreter Reference(Type type, string typeName)
 		{
-			_settings.Reference(type, typeName);
-			return this;
+			return Reference(new ReferenceType(typeName, type));
 		}
 
 		/// <summary>
@@ -241,13 +327,21 @@ namespace DynamicExpresso
 		/// <returns></returns>
 		public Interpreter Reference(ReferenceType type)
 		{
-			_settings.Reference(type);
+			if (type == null)
+				throw new ArgumentNullException(nameof(type));
+
+			_settings.KnownTypes[type.Name] = type;
+
+			foreach (var extensionMethod in type.ExtensionMethods)
+			{
+				_settings.ExtensionMethods.Add(extensionMethod);
+			}
+
 			return this;
 		}
 		#endregion
 
 		#region Parse
-
 		/// <summary>
 		/// Parse a text expression and returns a Lambda class that can be used to invoke it.
 		/// </summary>
@@ -326,7 +420,7 @@ namespace DynamicExpresso
 
 		public Lambda ParseAs<TDelegate>(string expressionText, params string[] parametersNames)
 		{
-			return ParseAs(typeof(TDelegate), expressionText, parametersNames);
+			return ParseAs(typeof(TDelegate), expressionText, parametersNames); 
 		}
 
 		internal Lambda ParseAs(Type delegateType, string expressionText, params string[] parametersNames)
@@ -376,7 +470,7 @@ namespace DynamicExpresso
 		#region Detection
 		public IdentifiersInfo DetectIdentifiers(string expression)
 		{
-			var detector = new Detector(_settings.ParserSettings);
+			var detector = new Detector(_settings);
 
 			return detector.DetectIdentifiers(expression);
 		}
@@ -388,7 +482,7 @@ namespace DynamicExpresso
 		{
 			var arguments = new ParserArguments(
 												expressionText,
-												_settings.ParserSettings,
+												_settings,
 												expressionType,
 												parameters);
